+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased ([main])
 
+### Fixed
+- Fix rendering of goals panel when no proof is active in Coq >= 8.16
+
 ## [1.7.1]
 
 ### Added